fix(iad-640): added test fixtures for position batch handler
