+++ package.json
@@ -54,6 +54,7 @@
     "test:int": "npx tape 'test/integration/**/*.test.js' ",
     "test:int-override": "npx tape 'test/integration-override/**/*.test.js'",
     "test:int:spec": "npm run test:int | npx tap-spec",
+    "test:int:handler-pos-batch": "npx tape 'test/integration-override/handlers/positions/handlerBatch.test.js'",
     "test:xint": "npm run test:int | tee /dev/tty | tap-xunit > ./test/results/xunit-integration.xml",
     "test:xint-override": "npm run test:int-override | tee /dev/tty | tap-xunit > ./test/results/xunit-integration-override.xml",
     "test:integration": "CSL_LOG_LEVEL=debug ./test/scripts/test-integration.sh",
@@ -163,11 +164,11 @@
       "ajv": "6.12.3",
       "ejs": "3.1.10",
       "sanitize-html": "2.12.1",
-      "markdown-it": "12.3.2"
+      "markdown-it": "14.1.1"
     },
     "widdershins": {
       "swagger2openapi": "7.0.8",
-      "markdown-it": "12.3.2"
+      "markdown-it": "14.1.1"
     },
     "hapi-auth-bearer-token": {
       "joi": "18.0.1"
@@ -192,7 +193,8 @@
   },
   "standard": {
     "ignore": [
-      "/migrations/**"
+      "/migrations/**",
+      "/docker/**"
     ]
   },
   "standard-version": {
